move `parseEntityYaml` to the node package
