+++ plugins/catalog-backend/src/index.ts
@@ -25,4 +25,3 @@ export * from './processors';
 export * from './processing';
 export * from './service';
 export * from './constants';
-export * from './util';
